requirements.txt
anyio==3.5.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bleach==5.0.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
cloudinary==1.29.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.1.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-debug-toolbar==3.6.0
django-filter==22.1
django-heroku==0.3.1
django-rest-registration==0.7.2
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
entrypoints==0.3
eth-hash==0.3.2
eth-typing==2.3.0
excel==1.0.0
executing==0.8.2
fastapi==0.73.0
fastjsonschema==2.15.3
filelock==3.4.2
Flask==2.0.3
fonttools==4.28.5
gunicorn==20.0.4
h11==0.13.0
html5lib==1.1
idna==2.8
inflection==0.5.1
ipykernel==6.7.0
ipython==8.0.1
itsdangerous==2.1.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
json2==0.4.0
json2json==0.1.0
JSON4JSON==0.4.3
jsonify==0.5
jsonschema==4.6.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.2.2
kiwisolver==1.3.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mplfinance==0.12.8b6
msgpack==0.6.2
mutagen==1.45.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.4
numpy==1.22.1
oauthlib==3.2.0
openapi-codec==1.3.2
outcome==1.1.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2022.9.4
platformdirs==2.4.1
progress==1.5
prompt-toolkit==3.0.24
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
pykalman==0.9.5
pyOpenSSL==21.0.0
pyparsing==2.4.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.25
python-magic-bin==0.4.14
python3-openid==3.2.0
pytoml==0.1.21
pytube==12.1.0
pytz==2022.2.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.0
simplejson==3.17.6
six==1.14.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.1
stack-data==0.1.4
starlette==0.17.1
stripe==2.58.0
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tpqoa==0.0.55
traitlets==5.2.2.post1
trio==0.19.0
trio-websocket==0.9.2
twilio==7.5.1
typing_extensions==4.0.1
tzdata==2021.5
ujson==5.1.0
uritemplate==4.1.1
urllib3==1.26.8
v20==3.0.25.0
virtualenv==20.13.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==5.2.0
wsproto==1.0.0
xlrd
XlsxWriter
